requirements.txt
pyserial>=3.5
rich>=13.7.0
matplotlib>=3.8.2
numpy>=1.26.3 